requirements.txt
google-genai
pydantic
streamlit
fastapi
uvicorn
python-multipart
